workers/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis & NATS
redis==5.0.1
nats-py==2.6.0

# AWS/MinIO
boto3==1.34.0
minio==7.2.0

# Script parsing
PyPDF2==3.0.1
python-docx==1.1.0
lxml==4.9.3
beautifulsoup4==4.12.2
fountain==0.3.0

# AI/ML libraries
torch==2.1.1
transformers==4.36.0
diffusers==0.24.0
accelerate==0.25.0
safetensors==0.4.1

# Media processing
ffmpeg-python==0.2.0
Pillow==10.1.0
opencv-python==4.8.1.78

# Observability & Monitoring
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-exporter-prometheus==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-resources==1.21.0
opentelemetry-semantic-conventions==0.42b0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Utilities
python-multipart==0.0.6
httpx==0.25.2
tenacity==8.2.3
python-json-logger==2.0.7
